requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
pymysql
cryptography
